put classes where they belong
